Move RunningNorm tests to networks now it ahs been factored out
