+++ tests/util/test_networks.py
@@ -1,0 +1,89 @@
+"""Tests `imitation.util.networks`."""
+
+import pytest
+import torch as th
+
+from imitation.util import networks
+
+
+def test_running_norm_identity() -> None:
+    """Tests running norm starts and stays at identity function.
+
+    Specifically, we test in evaluation mode (initializatn should not change)
+    and in training mode with already normalized data.
+    """
+    running_norm = networks.RunningNorm(1, eps=0.0)
+    x = th.Tensor([-1.0, 0.0, 7.32, 42.0])
+    running_norm.eval()  # stats should not change in eval mode
+    for i in range(10):
+        th.testing.assert_equal(running_norm.forward(x), x)
+    running_norm.train()  # stats will change in eval mode
+    normalized = th.Tensor([-1, 1])  # mean 0, variance 1
+    for i in range(10):
+        th.testing.assert_equal(running_norm.forward(normalized), normalized)
+
+
+def test_running_norm_eval_fixed(
+    batch_size: int = 8,
+    num_batches: int = 10,
+    num_features: int = 4,
+) -> None:
+    """Tests that stats do not change when in eval mode and do when in training."""
+    running_norm = networks.RunningNorm(num_features)
+
+    def do_forward(shift: float = 0.0, scale: float = 1.0):
+        for i in range(num_batches):
+            data = th.rand(batch_size, num_features) * scale + shift
+            running_norm.forward(data)
+
+    with th.random.fork_rng():
+        th.random.manual_seed(42)
+
+        do_forward()
+        current_mean = th.clone(running_norm.running_mean)
+        current_var = th.clone(running_norm.running_var)
+
+        running_norm.eval()
+        do_forward()
+        th.testing.assert_equal(running_norm.running_mean, current_mean)
+        th.testing.assert_equal(running_norm.running_var, current_var)
+
+        running_norm.train()
+        do_forward(1.0, 2.0)
+        assert th.all((running_norm.running_mean - current_mean).abs() > 0.01)
+        assert th.all((running_norm.running_var - current_var).abs() > 0.01)
+
+
+@pytest.mark.parametrize("batch_size", [1, 8])
+def test_running_norm_matches_dist(batch_size: int) -> None:
+    """Test running norm converges to empirical distribution."""
+    mean = th.Tensor([-1.3, 0.0, 42])
+    var = th.Tensor([0.1, 1.0, 42])
+    sd = th.sqrt(var)
+
+    num_dims = len(mean)
+    running_norm = networks.RunningNorm(num_dims)
+    running_norm.train()
+
+    num_samples = 256
+    with th.random.fork_rng():
+        th.random.manual_seed(42)
+        data = th.randn(num_samples, num_dims) * sd + mean
+        for start in range(0, num_samples, batch_size):
+            batch = data[start : start + batch_size]
+            running_norm.forward(batch)
+
+    empirical_mean = th.mean(data, dim=0)
+    empirical_var = th.var(data, dim=0, unbiased=False)
+
+    normalized = th.Tensor([[-1.0], [0.0], [1.0], [42.0]])
+    normalized = th.tile(normalized, (1, 3))
+    scaled = normalized * th.sqrt(empirical_var + running_norm.eps) + empirical_mean
+    running_norm.eval()
+    for i in range(5):
+        th.testing.assert_close(running_norm.forward(scaled), normalized)
+
+    # Stats should match empirical mean (and be unchanged by eval)
+    th.testing.assert_close(running_norm.running_mean, empirical_mean)
+    th.testing.assert_close(running_norm.running_var, empirical_var)
+    assert running_norm.count == num_samples
